Split each widget into a ProverKey and VerifierKey
