+++ src/proof_system/mod.rs
@@ -1,12 +1,16 @@
 //! proving system
 
 pub(crate) mod linearisation_poly;
-pub(crate) mod preprocessed_circuit;
-pub(crate) mod proof;
+/// Represents the Prover and Verifier key for a PLONK Circuit
+pub mod preprocessed_circuit;
+/// Represents a PLONK Proof
+pub mod proof;
 pub(crate) mod proof_system_errors;
-pub(crate) mod prover;
+/// Represents a PLONK Prover
+pub mod prover;
 pub(crate) mod quotient_poly;
-pub(crate) mod verifier;
+/// Represents a PLONK Verifier
+pub mod verifier;
 pub(crate) mod widget;
 
 pub use preprocessed_circuit::PreProcessedCircuit;
